Implement tests for SSH key checks

+++ dokkupy/dokku_cli.py
@@ -21,17 +21,18 @@ class Dokku:
     ):
         self._ssh_prefix = []
         self.__files_to_delete = []
-        self._local_user = getpass.getuser()
+        self.local_user = getpass.getuser()
+        self.ssh_host, self.ssh_port, self.ssh_user = None, None, None
         if ssh_host:
             self.ssh_host, self.ssh_port, self.ssh_user = ssh_host, ssh_port, ssh_user
             self.ssh_private_key = (
                 Path(ssh_private_key).expanduser().absolute() if ssh_private_key is not None else None
             )
             if ssh_private_key is None:
-                raise ValueError("ssh_private_key must be provided to ensure the execution is non-interactive")
+                raise ValueError("`ssh_private_key` must be provided to ensure the execution is non-interactive")
             elif ssh.key_requires_password(self.ssh_private_key):
                 if ssh_key_password is None:
-                    raise ValueError("The SSH key password must be provided so the execution is non-interactive")
+                    raise ValueError("`ssh_key_password` must be provided so the execution is non-interactive")
                 self.ssh_private_key = ssh.key_unlock(self.ssh_private_key, ssh_key_password)
                 self.__files_to_delete.append(self.ssh_private_key)
             self._ssh_prefix = ssh.command(
@@ -40,6 +41,7 @@ class Dokku:
                 port=self.ssh_port,
                 private_key=self.ssh_private_key,
             )
+        self.via_ssh = len(self._ssh_prefix) > 0
 
         # Instantiate default plugins
         # TODO: autodiscover based on `DokkuPlugin` subclasses?
@@ -58,18 +60,18 @@ class Dokku:
             self.plugins[name] = instance
 
     def __del__(self):
-        for filename in self.__files_to_delete:
-            if filename.exists():
-                filename.unlink()
+        if hasattr(self, "_Dokku__files_to_delete"):
+            for filename in self.__files_to_delete:
+                if filename.exists():
+                    filename.unlink()
 
     def _prepare_command(self, command: Command) -> tuple[str]:
         """Prepare the final command to be executed, considering sudo, local/remote user and the command itself"""
         cmd = list(command.command)
         use_sudo = command.sudo
-        execute_via_ssh = len(self._ssh_prefix) > 0
         is_dokku_command = cmd[0] == "dokku"
 
-        if execute_via_ssh:  # May consider: self.ssh_user, use_sudo, is_dokku_command. Don't care: self._local_user
+        if self.via_ssh:  # May consider: self.ssh_user, use_sudo, is_dokku_command. Don't care: self.local_user
             if self.ssh_user == "dokku":
                 # If running via SSH and the remote user is `dokku`, the only commands we can execute are Dokku
                 # commands
@@ -81,8 +83,8 @@ class Dokku:
                     cmd = cmd[1:]  # The `dokku` command is not passed via SSH
             elif self.ssh_user == "root":
                 use_sudo = False  # If running via SSH and the remote user is `root`, `sudo` is not needed
-        else:  # May consider: self._local_user, use_sudo, is_dokku_command. Don't care: self.ssh_user
-            if self._local_user == "root":
+        else:  # May consider: self.local_user, use_sudo, is_dokku_command. Don't care: self.ssh_user
+            if self.local_user == "root":
                 use_sudo = False  # If executing locally and the local user is `root`, `sudo` is not needed
         return self._ssh_prefix + (["sudo"] if use_sudo else []) + cmd
